requirements.txt
asgiref==3.10.0
Django==5.2.8
djangorestframework==3.16.1
iniconfig==2.3.0
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
pytest==9.0.1
pytest-django==4.11.1
sqlparse==0.5.3
